Move HereFunctionHandler to its own file and store location string Launch activity in EntitySelectActivity and store result in hereFunctionHandler
