+++ app/src/org/commcare/android/models/NodeEntityFactory.java
@@ -1,16 +1,16 @@
 package org.commcare.android.models;
 
+import org.commcare.dalvik.activities.EntitySelectActivity;
+import org.commcare.dalvik.geo.HereFunctionHandler;
 import org.commcare.suite.model.Detail;
 import org.commcare.suite.model.DetailField;
 import org.commcare.suite.model.Text;
 import org.javarosa.core.model.condition.EvaluationContext;
-import org.javarosa.core.model.condition.IFunctionHandler;
 import org.javarosa.core.model.instance.TreeReference;
 import org.javarosa.xpath.XPathException;
 import org.javarosa.xpath.parser.XPathSyntaxException;
 
 import java.util.List;
-import java.util.Vector;
 
 /**
  * @author ctsims
@@ -26,7 +26,7 @@ public class NodeEntityFactory {
     public NodeEntityFactory(Detail d, EvaluationContext ec) {
         this.detail = d;
         this.ec = ec;
-        this.hereFunctionHandler = new HereFunctionHandler();
+        this.hereFunctionHandler = EntitySelectActivity.hereFunctionHandler;
     }
 
     public Detail getDetail() {
@@ -123,34 +123,6 @@ public class NodeEntityFactory {
                 throw new RuntimeException("A Node Entity Factory was not prepared before usage. prepareEntities() must be called before a call to isEntitySetReady()");
             }
             return isEntitySetReadyInternal();
-        }
-    }
-
-    private class HereFunctionHandler implements IFunctionHandler {
-        public static final String HERE_NAME = "here";
-
-        public HereFunctionHandler() {}
-
-        public String getName() {
-            return HERE_NAME;
-        }
-
-        public Vector getPrototypes() {
-            Vector p = new Vector();
-            p.addElement(new Class[0]);
-            return p;
-        }
-
-        public boolean rawArgs() {
-            return false;
-        }
-
-        public boolean realTime() {
-            return true;
-        }
-
-        public Object eval(Object[] args, EvaluationContext ec) {
-            return "42.33 -71.11";
         }
     }
 }
